Make the test_app functional.
The test_app isn't just for testing any more. It can now run as a Django project locally, or be pushed to Heroku, to test real, remote email receiving - which is difficult to replicate with tests.

From settings.py:

"""
The test_app project is primarily used for running the django_inbound_email tests within a Django project, however, this settings file does allow for a secondary set of 'real-world' tests. If the environment settings allow, the test_app will use the django_inbound_email app to listen for incoming messages, which it will bounceback to the sender.

In addition, the top-level requirements.txt and Procfile files allow for the project to be deployed, as-is, to Heroku for real-world testing.

If you have an Heroku account, it's as simple as:

    # pull the latest source and cd into the directory
    $ git pull git@github.com:yunojuno/django-inbound-email.git
    $ cd django-inbound-email
    # create a new Heroku app, set the parser to use.
    django-inbound-email$ heroku create
    django-inbound-email$ heroku config:set INBOUND_EMAIL_PARSER=path.to.parser
    # push latest master to new Heroku app
    django-inbound-email$ git push heroku master:master

If the app created by Heroku is called "safe-earth-8826", then you will now have an inbound handler on http://safe-earth-8826.herokuapp.com/emails/inbound/
You can either POST direct to this handler (e.g. using the Postman chrome app), or even configure your email handler and try some real tests.

NB It uses the Django dev server to serve the requests - I hope I don't need to point out that this is in no way sufficient to run a production application. You run this app at your own risk.