Remove support for Django < 1.8, add support for Django 1.9, 1.10
I’ve updated tox.ini to run the tests against 1.8, 1.9, 1.10, but looking at it it
could do with a major refactor. Lots learnt about redistributable packages
since this went live.

Have bumped the version to v0.9.0 as it’s a breaking change - the package
name is now `inbound_email` without the django prefix.